Create Tree class
Tree class takes an array as an argument and uses
Node to build itself.

mergeSort() sorts the array on first iteration of
buildTree()

buildTree() takes the middle element of an array
as the root Node and builds the tree of Nodes on
it.

prettyPrint() prints the tree in console.

index.mjs shows the classes in use.